+++ frontend/src/pages/host/create/funnel/SpaceCreateFunnel.styles.ts
