+++ frontend/src/pages/host/create/funnelElements/SpaceDescriptionElement.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import TextareaInput from '../../../components/@common/inputs/textareaInput/TextareaInput';
-import { CONSTRAINTS } from '../../../constants/constraints';
-import type { FunnelElementProps } from '../../../types/funnel.type';
-import { calculateValidLength } from '../../../utils/grapheme';
-import { createErrorMessageWithValidators } from '../../../validators/createErrorMessageWithValidators';
+import TextareaInput from '../../../../components/@common/inputs/textareaInput/TextareaInput';
+import { CONSTRAINTS } from '../../../../constants/constraints';
+import type { FunnelElementProps } from '../../../../types/funnel.type';
+import { calculateValidLength } from '../../../../utils/grapheme';
+import { createErrorMessageWithValidators } from '../../../../validators/createErrorMessageWithValidators';
 import { funnelValidators } from '../funnel/funnel.validators';
 import FunnelBasePage from '../funnel/funnelBasePage/FunnelBasePage';
