+++ frontend/src/pages/host/create/funnelElements/SpaceNameElement.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import TextInput from '../../../components/@common/inputs/textInput/TextInput';
-import { CONSTRAINTS } from '../../../constants/constraints';
-import type { FunnelElementProps } from '../../../types/funnel.type';
-import { calculateValidLength } from '../../../utils/grapheme';
-import { createErrorMessageWithValidators } from '../../../validators/createErrorMessageWithValidators';
+import TextInput from '../../../../components/@common/inputs/textInput/TextInput';
+import { CONSTRAINTS } from '../../../../constants/constraints';
+import type { FunnelElementProps } from '../../../../types/funnel.type';
+import { calculateValidLength } from '../../../../utils/grapheme';
+import { createErrorMessageWithValidators } from '../../../../validators/createErrorMessageWithValidators';
 import { funnelValidators } from '../funnel/funnel.validators';
 import FunnelBasePage from '../funnel/funnelBasePage/FunnelBasePage';
