+++ frontend/src/pages/host/create/funnelElements/spaceCheckElement/SpaceCheckElement.styles.ts
