+++ frontend/src/pages/host/create/funnelElements/spaceDetailElement/SpaceDetailElement.styles.tsx
