+++ frontend/src/pages/host/create/funnelElements/spaceDetailElement/SpaceDetailElement.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
-import TextInput from '../../../../components/@common/inputs/textInput/TextInput';
-import PhotoUploadButton from '../../../../components/specific/photoUploadButton/PhotoUploadButton';
-import useLocalFile from '../../../../hooks/domain/useLocaleFile';
+import TextInput from '../../../../../components/@common/inputs/textInput/TextInput';
+import PhotoUploadButton from '../../../../../components/specific/photoUploadButton/PhotoUploadButton';
+import useLocalFile from '../../../../../hooks/domain/useLocaleFile';
 import type {
   FunnelElementProps,
   SpaceDetailElementInfos,
-} from '../../../../types/funnel.type';
-import { createErrorMessageWithValidators } from '../../../../validators/createErrorMessageWithValidators';
+} from '../../../../../types/funnel.type';
+import { createErrorMessageWithValidators } from '../../../../../validators/createErrorMessageWithValidators';
 import { funnelValidators } from '../../funnel/funnel.validators';
 import FunnelBasePage from '../../funnel/funnelBasePage/FunnelBasePage';
 import * as S from './SpaceDetailElement.styles';
